requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
adwaita-icon-theme=40.1.1=hca03da5_1
anyio=3.5.0=py39hca03da5_0
appnope=0.1.2=py39hca03da5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39h1a28f6b_0
asttokens=2.0.5=pyhd3eb1b0_0
atk-1.0=2.36.0=h7fe96df_0
attrs=21.4.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py39hca03da5_0
blas=1.0=openblas
bleach=4.1.0=pyhd3eb1b0_0
bottleneck=1.3.4=py39heec5a64_0
brotli=1.0.7=hc377ac9_0
brotlipy=0.7.0=py39h1a28f6b_1002
ca-certificates=2022.4.26=hca03da5_0
cairo=1.16.0=h29d4eff_2
certifi=2022.6.15=py39hca03da5_0
cffi=1.15.0=py39h22df2f2_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39hca03da5_0
cryptography=37.0.1=py39h834c97f_0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=h8280c03_0
debugpy=1.5.1=py39hc377ac9_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py39hca03da5_0
epoxy=1.5.4=h1a28f6b_2
executing=0.8.3=pyhd3eb1b0_0
expat=2.4.4=hc377ac9_0
filelock=3.6.0=pyhd3eb1b0_0
fontconfig=2.13.1=h1f4a5ec_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h1192e45_0
fribidi=1.0.10=h1a28f6b_0
future=0.18.2=py39hca03da5_1
gdk-pixbuf=2.42.8=hb4bb6b2_0
gettext=0.21.0=h826f4ad_0
giflib=5.2.1=h1a28f6b_0
glib=2.69.1=h98b2900_1
gobject-introspection=1.72.0=py39h18a2b1f_0
graphite2=1.3.14=hc377ac9_1
gtk3=3.24.21=h321fa23_1
harfbuzz=2.8.1=hffeda63_0
hicolor-icon-theme=0.17=hca03da5_2
huggingface_hub=0.2.1=pyhd3eb1b0_0
icu=68.1=hc377ac9_0
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py39hca03da5_0
importlib_metadata=4.11.3=hd3eb1b0_0
ipykernel=6.9.1=py39hca03da5_0
ipython=8.3.0=py39hca03da5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd3eb1b0_1
jedi=0.18.1=py39hca03da5_1
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h1a28f6b_0
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.4.0=py39hca03da5_0
jupyter=1.0.0=py39hca03da5_7
jupyter_client=7.2.2=py39hca03da5_0
jupyter_console=6.4.3=pyhd3eb1b0_0
jupyter_core=4.10.0=py39hca03da5_0
jupyter_server=1.17.1=py39hca03da5_0
jupyterlab=3.3.2=pyhd3eb1b0_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.12.0=py39hca03da5_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
kiwisolver=1.4.2=py39hc377ac9_0
krb5=1.19.2=h3b8d789_0
lcms2=2.12=hba8e193_0
libcxx=12.0.0=hf6beb65_1
libedit=3.1.20210910=h1a28f6b_0
libevent=2.1.12=hf27765b_0
libffi=3.4.2=hc377ac9_4
libgfortran=5.0.0=11_2_0_he6877d6_26
libgfortran5=11.2.0=he6877d6_26
libiconv=1.16=h1a28f6b_2
libopenblas=0.3.20=hea475bc_0
libpng=1.6.37=hb8d0fd4_0
libpq=12.9=h65cfe13_3
librsvg=2.50.7=h90a7944_0
libsodium=1.0.18=h1a28f6b_0
libtiff=4.2.0=h01837e1_1
libwebp=1.2.2=h68602c7_0
libwebp-base=1.2.2=h1a28f6b_0
libxml2=2.9.14=h8c5e841_0
libxslt=1.1.35=h9833966_0
llvm-openmp=12.0.0=haf9daa7_1
lz4-c=1.9.3=hc377ac9_0
markupsafe=2.1.1=py39h1a28f6b_0
matplotlib=3.5.1=py39hca03da5_1
matplotlib-base=3.5.1=py39hc377ac9_1
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h1a28f6b_1000
munkres=1.1.4=py_0
nbclassic=0.3.5=pyhd3eb1b0_0
nbclient=0.5.13=py39hca03da5_0
nbconvert=6.4.4=py39hca03da5_0
nbformat=5.3.0=py39hca03da5_0
ncurses=6.3=h1a28f6b_2
nest-asyncio=1.5.5=py39hca03da5_0
networkx=2.7.1=pyhd3eb1b0_0
ninja=1.10.2=hca03da5_5
ninja-base=1.10.2=h525c30c_5
nltk=3.7=pyhd3eb1b0_0
notebook=6.4.11=py39hca03da5_0
nspr=4.33=hc377ac9_0
nss=3.74=h142855e_0
numexpr=2.8.1=py39h144ceef_2
numpy=1.22.3=py39h25ab29e_0
numpy-base=1.22.3=py39h974a1f5_0
openssl=1.1.1p=h1a28f6b_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py39hc377ac9_0
pandocfilters=1.5.0=pyhd3eb1b0_0
pango=1.48.7=h79d33a7_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=hc377ac9_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py39h4d1bdd5_0
pip=21.2.4=py39hca03da5_0
pixman=0.40.0=h1a28f6b_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
prompt_toolkit=3.0.20=hd3eb1b0_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.15.2=py39he8f2410_0
pyqt5-sip=4.19.25=pypi_0
pyqtchart=5.15.2=pypi_0
pyrsistent=0.18.0=py39h1a28f6b_0
pysocks=1.7.1=py39hca03da5_0
python=3.9.12=hbdb9e5c_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
pytorch=1.13.0.dev20220629=py3.9_0
pytz=2022.1=py39hca03da5_0
pyyaml=6.0=py39h1a28f6b_0
pyzmq=22.3.0=py39hc377ac9_2
qt=5.15.2=h4e944ae_2
qtconsole=5.3.0=pyhd3eb1b0_0
qtpy=2.0.1=pyhd3eb1b0_0
readline=8.1.2=h1a28f6b_1
regex=2022.3.15=py39h1a28f6b_0
requests=2.28.0=py39hca03da5_0
sacremoses=0.0.43=pyhd3eb1b0_0
scikit-learn=1.0.2=py39h9197a36_1
scipy=1.7.3=py39h2f0f56f_0
seaborn=0.11.2=pyhd3eb1b0_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=61.2.0=py39hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=py39hca03da5_1
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.38.5=h1058600_0
stack_data=0.2.0=pyhd3eb1b0_0
terminado=0.13.1=py39hca03da5_0
testpath=0.6.0=py39hca03da5_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=hb8d0fd4_0
tokenizers=0.11.4=py39h5c5695e_1
tornado=6.1=py39h1a28f6b_0
tqdm=4.64.0=py39hca03da5_0
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.18.0=py39hca03da5_0
typing-extensions=4.1.1=hd3eb1b0_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39hca03da5_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py39hca03da5_1
websocket-client=0.58.0=py39hca03da5_4
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=py39hca03da5_0
xz=5.2.5=h1a28f6b_1
yaml=0.2.5=h1a28f6b_0
zeromq=4.3.4=hc377ac9_0
zipp=3.8.0=py39hca03da5_0
zlib=1.2.12=h5a0b063_2
zstd=1.5.2=h8574219_0
